+++ apps/katalog/src/views/PercentileBandView.vue
@@ -6,7 +6,7 @@ import Card from 'primevue/card';
 import Skeleton from 'primevue/skeleton';
 import Message from 'primevue/message';
 import Tag from 'primevue/tag';
-import PercentileBandChart from '../components/PercentileBandChart.vue';
+import { Perzentilbaender as PercentileBandChart } from '@tba3/bausteine/vue';
 import ComponentDocs from '../components/ComponentDocs.vue';
 import { t } from '../i18n';